*** empty log message ***

+++ Abills/mysql/Acct.pm
@@ -93,24 +93,8 @@ if ($acct_status_type == 1) {
     WHERE user_name='$RAD->{USER_NAME}' and acct_session_id='$RAD->{ACCT_SESSION_ID}';");
     
   if ($self->{list}->[0]->[0] < 1) {
-    #Get TP_ID
-    $self->query($db, "SELECT u.uid, dv.tp_id, dv.join_service FROM (users u, dv_main dv)
-     WHERE u.uid=dv.uid and u.id='$RAD->{USER_NAME}';");
-    if ($self->{TOTAL} > 0) {
-      ($self->{UID},
-       $self->{TP_ID},
-       $self->{JOIN_SERVICE})= @{ $self->{list}->[0] };
-       
-       if ($self->{JOIN_SERVICE}) {
-       	 if ($self->{JOIN_SERVICE} == 1) {
-       	 	 $self->{JOIN_SERVICE}=$self->{UID};
-       	  }
-       	 $self->{TP_ID}='';
-        }
-     }
-    else {
-    	$RAD->{USER_NAME}='! '.$RAD->{USER_NAME};
-     }
+    $self->query($db, "SELECT count(uid) FROM dv_calls 
+      WHERE user_name='$RAD->{USER_NAME}' AND nas_id='$NAS->{NAS_ID}' AND acct_session_id='IP' AND framed_ip_address=INET_ATON('$RAD->{FRAMED_IP_ADDRESS}');");
     
     #Get connection speed 
     if ($RAD->{X_ASCEND_DATA_RATE} && $RAD->{X_ASCEND_XMIT_RATE}) {
@@ -119,26 +103,59 @@ if ($acct_status_type == 1) {
     elsif ($RAD->{CISCO_SERVICE_INFO}) {
       $RAD->{CONNECT_INFO}="$RAD->{CISCO_SERVICE_INFO}";
      }
+    
+    if ($self->{list}->[0]->[0] == 1) {
+      my $sql = "UPDATE dv_calls SET
+       status='$acct_status_type',
+       started=$SESSION_START, 
+       lupdated=UNIX_TIMESTAMP(), 
+       nas_port_id='$RAD->{NAS_PORT}', 
+       acct_session_id='$RAD->{ACCT_SESSION_ID}', 
+       CID='$RAD->{CALLING_STATION_ID}', 
+       CONNECT_INFO='$RAD->{CONNECT_INFO}',   
+       );";
+      $self->query($db, "$sql", 'do');
+     }
+    else {
+      #Get TP_ID
+      $self->query($db, "SELECT u.uid, dv.tp_id, dv.join_service FROM (users u, dv_main dv)
+       WHERE u.uid=dv.uid and u.id='$RAD->{USER_NAME}';");
+      if ($self->{TOTAL} > 0) {
+        ($self->{UID},
+         $self->{TP_ID},
+         $self->{JOIN_SERVICE})= @{ $self->{list}->[0] };
+       
+         if ($self->{JOIN_SERVICE}) {
+         	 if ($self->{JOIN_SERVICE} == 1) {
+         	 	 $self->{JOIN_SERVICE}=$self->{UID};
+         	  }
+       	   $self->{TP_ID}='';
+          }
+       }
+      else {
+    	  $RAD->{USER_NAME}='! '.$RAD->{USER_NAME};
+       }
 
-    my $sql = "INSERT INTO dv_calls
-     (status, user_name, started, lupdated, nas_ip_address, nas_port_id, acct_session_id, framed_ip_address, CID, CONNECT_INFO, nas_id, tp_id,
-      uid, join_service)
-       values ('$acct_status_type', 
-      '$RAD->{USER_NAME}', 
-      $SESSION_START, 
-      UNIX_TIMESTAMP(), 
-      INET_ATON('$RAD->{NAS_IP_ADDRESS}'),
-      '$RAD->{NAS_PORT}', 
-      '$RAD->{ACCT_SESSION_ID}',
-      INET_ATON('$RAD->{FRAMED_IP_ADDRESS}'), 
-      '$RAD->{CALLING_STATION_ID}', 
-      '$RAD->{CONNECT_INFO}', 
-      '$NAS->{NAS_ID}',
-      '$self->{TP_ID}', '$self->{UID}',
-      '$self->{JOIN_SERVICE}');";
-    $self->query($db, "$sql", 'do');
+      my $sql = "INSERT INTO dv_calls
+       (status, user_name, started, lupdated, nas_ip_address, nas_port_id, acct_session_id, framed_ip_address, CID, CONNECT_INFO,   nas_id, tp_id,
+        uid, join_service)
+         values ('$acct_status_type', 
+        '$RAD->{USER_NAME}', 
+        $SESSION_START, 
+        UNIX_TIMESTAMP(), 
+        INET_ATON('$RAD->{NAS_IP_ADDRESS}'),
+        '$RAD->{NAS_PORT}', 
+        '$RAD->{ACCT_SESSION_ID}',
+        INET_ATON('$RAD->{FRAMED_IP_ADDRESS}'), 
+        '$RAD->{CALLING_STATION_ID}', 
+        '$RAD->{CONNECT_INFO}', 
+        '$NAS->{NAS_ID}',
+        '$self->{TP_ID}', '$self->{UID}',
+        '$self->{JOIN_SERVICE}');";
+      $self->query($db, "$sql", 'do');
 
-    $self->query($db, "DELETE FROM dv_calls WHERE nas_id='$NAS->{NAS_ID}' AND acct_session_id='IP' AND (framed_ip_address=INET_ATON('$RAD->{FRAMED_IP_ADDRESS}') or UNIX_TIMESTAMP()-UNIX_TIMESTAMP(started) > 120 );", 'do');
+      $self->query($db, "DELETE FROM dv_calls WHERE nas_id='$NAS->{NAS_ID}' AND acct_session_id='IP' AND (framed_ip_address=INET_ATON('$RAD->{FRAMED_IP_ADDRESS}') or UNIX_TIMESTAMP()-UNIX_TIMESTAMP(started) > 120 );", 'do');
+     }
   }
  }
 # Stop status
